supercat_automation_agent/supercat_automation/requirements.txt
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
typing-extensions==4.14.1
beautifulsoup4==4.12.2
playwright==1.40.0
requests==2.31.0
cloudscraper==1.2.71
2captcha-python==1.1.1
supabase==2.4.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.0
pandas==2.1.3
numpy==1.24.3
openpyxl==3.1.2
openai==1.3.0
httpx==0.27.0
aiohttp==3.9.1
googlemaps==4.10.0
schedule==1.2.1
celery==5.3.6
redis==5.0.1
streamlit==1.28.2
plotly==5.18.0
altair==5.2.0
loguru==0.7.2
sentry-sdk==1.38.0
pytest==7.4.3
pytest-asyncio==0.21.1
Faker==20.1.0
python-dateutil==2.8.2
pyyaml==6.0.1
jinja2==3.1.2
phonenumbers==8.13.26
email-validator==2.1.0
flask==2.3.3
